move execution pending envelolpe logic to its own file
